test: split schema tests
